requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
attrs==25.1.0
blinker==1.8.2
cachetools==5.5.1
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.1.0
comm==0.2.2
contourpy==1.3.1
currencyapicom==0.1.1
cycler==0.12.1
dask==2024.10.0
debugpy==1.8.8
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
Django==5.1
djangorestframework==3.15.2
everapi==0.1.1
executing==2.1.0
Flask==3.0.3
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.12
GitPython==3.1.44
gotrue==2.11.3
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.7
instaloader==4.14.1
ipykernel==6.29.5
ipython==8.29.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
locket==1.0.0
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.1.0
narwhals==1.24.1
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.0.0
packaging==24.1
panda==0.3.1
pandas==2.2.2
parso==0.8.4
partd==1.4.2
pexpect==4.9.0
pikepdf==9.0.0
pillow==10.4.0
platformdirs==4.3.6
postgrest==0.19.3
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.0
PyBluez @ git+https://github.com/pybluez/pybluez@82cbba8a1ebd4c1e3442dfafd8581d58c50fa39e
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.18.0
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyparsing==3.2.1
python-constraint==1.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
realtime==2.3.0
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
scikit-learn==1.6.1
scipy==1.15.1
setuptools==70.2.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
sqlparse==0.5.1
stack-data==0.6.3
storage3==0.11.3
streamlit==1.41.1
StrEnum==0.4.15
supabase==2.12.0
supafunc==0.9.3
tenacity==9.0.0
threadpoolctl==3.5.0
toml==0.10.2
toolz==1.0.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
websockets==14.2
Werkzeug==3.0.3
wheel==0.44.0
wrapt==1.16.0
yarl==1.18.3
